Initial implementation of an rsync Tasker plugin
Import the tasker plugin library and android-rsync binaries library and
put together a very bare-bones implementation of an rsync plugin for
tasker.  This takes a single argument of the command line arguments
(with over-simplified argument splitting for the moment) and returns the
standard error and standard output produced by the process.