Switch Android lint warnings to errors and fix
Set warningsAsErrors true in lintOptions for the app so that lint
warnings cause the lint task to fail.  Set the details to be output to
standard out to save having to open the HTML report to see the details.

Disable NonConstantResourceId as this would need to be a change to the
Tasker plugin library (and it is rather unclear if this change is
actually landing in the AGP 7.0).

Fix the rest of the issues found by the lint task.