Include Dropbear binaries and add dbclient plugin
Import the android-dropbear binaries library (excluding the unneeded
dropbear server binary from being included in the app).

Add a dbclient plugin which is pretty much a straight copy of the rsync
plugin for the moment.  It seems to be a bit difficult to reduce
duplication between the various classes needed for each plugin (due to
the strong typing), so only repurpose RsyncOutput to a more generic
CommandOutput class for the time being.