Add a plugin action for generating a private key
A new plugin action to generate an SSH private key using dropbearkey.
The configuration activity defaults to Ed25519 but also allows selecting
ECDSA or RSA if required.  384/521 bit key sizes can be selected for
ECDSA and 2048/4096 bit for RSA.

By default an existing private key won't be overwritten if it already
exists, but there is a configuration option to bypass this if desired.