Load key by default for dbclient/rsync actions
Set up the correct environment for dbclient so that it can load the
private key when it looks for ~/.ssh/id_dropbear.  This required
updating to the 2020.81.1 release of android-dropbear as it includes a
patch which makes dbclient use HOME in preference to the value returned
by getpwuid()->pw_dir (which is hard-coded to always return "/data" for
Android apps).

Put the code for setting up the correct temporary 'HOME' directory with
the key as well as setting the HOME environment var and cleaning up
afterwards in a new ProcessEnv class.  Switch from Runtime.exec() to
ProcessBuilder as this gives a much cleaner interface to the process
environment.