Bump minSdkVersion from 16 to 19
The latest version of the Android NDK (r23) mentions in the release
notes that it is the last release supporting API levels 16, 17 & 18.  As
this will mean that the rsync and dropbear binaries will no-longer
support these API levels within the next year or so, pre-empt this by
upping minSdkVersion from 16 (Android 4.1) to 19 (Android 4.4).