Get app versionCode and versionName from git
Get the app's versionCode and versionName at build time from git
commands instead of having them hardcoded.  Set versionCode to the
total number of git commits and versionName to the output of git
describe (minus the odd 'g' prefix to the abbreviated object name if
applicable).

Update the Release action to fetch all commits during checkout so that
the versionCode is set to the correct value.