Declare permissions to read/write shared storage
For Android 10 and below, declare the WRITE_EXTERNAL_STORAGE permission.
This allows both reading from and writing to shared storage (with the
addition of requestLegacyExternalStorage=true just for Android 10).

For Android 11 and above, declare the MANAGE_EXTERNAL_STORAGE
permission.  This will require additional approval through the Play
Store but is the only way that the app can read and write to parts of
the shared storage which are outside the app's defined storage path.

Add an instrumented test that both reading from and writing to the
primary shared storage is successful and run the tests under Android 4.4
and 11 in addition to Android 10 on GitHub Actions.