Create & apply app icon and set theme colours
Design an icon and use Asset Studio to apply it to the app and generate
the various density PNG versions for pre-Android 8.0 devices.  Also
design a 'feature graphic' for the play store based on the icon.

Also use the colours from the icon as the basis for the app theme
colours.