Upgrade to AGP 7.1 and Gradle 7.2
Upgrade to Android Gradle Plugin 7.1.  This requires Gradle 7.2 or above
so upgrade this too.

The lint configuration has changed slightly for 7.2, so make slight
adjustments to cater for this while keeping the behaviour the same.

There is also a new DataExtractionRules lint check which pointed out
that allowBackup="false" is ignored for Android 12 and above, so add a
dataExtractionRules attribute pointing to an XML resource with just
blank exclude rules in the hope that this will exclude all data from
being backed up.  Also add an ignore attribute to prevent a lint warning
about also including the earlier format of this resource
(allowBackup="false" does this job instead).