Notify, download and prompt to install updates
When the rsync or dropbear actions are run spawn a background thread to
check for updates.  Cache the version information fetched from
nerdoftheherd.com for a week to limit data usage.  If an update is
available, show a low priority (e.g. silent) notification.

If the notification is tapped, show a new activity which allows the user
to view the more information URL or install the update.  When choosing
to install the update, the updated apk is downloaded and the user is
prompted to install the update.

For the VersionInfo and Version classes which handle deserialising,
caching and version number comparison etc, add a bunch of unit tests to
try and make this reasonably robust.