Allow specifying known SSH host keys
Add a field to the Dbclient and Rsync config activities to allow
specifying the known host keys for that action.  Extend ProcessEnv to
take this as an additional parameter and write the content to the
~/.ssh/known_hosts file in the generated home directory for that action.

This allows connecting to servers without having to use the dbclient -y
'Always accept hostkeys if they are unknown' switch and running the risk
of falling foul of MITM attacks.

Resolves issue #50.