Update targetSdk to 33 - update check now optional
Update the targetSdk from 32 to 33 and re-enable the OldTargetApi lint
check.  As part of targeting API level 33, permission must now be sought
from the user before most notifications appear.  Those shown by the
taskerpluginlibrary as part of running the foreground services will
continue to be shown but the update notification is affected by the
permission change.

To handle this, expose the 'checkForUpdates' parameter (previously just
used internally when running instrumented tests) to the user and as a
field for Tasker.  When the user exits the rsync or dbclient config
activities if; this is selected, the device is new enough to support the
notification permission and the permission is not already granted,
request the notification permission.  If the permission is granted, exit
the config activity or if not, unselect the update check parameter.