Update activity: Compare current ver with update
After the app has been updated it is still possible to navigate to the
Update activity sometimes which continues to claim that an update is
available.  As it is passed VersionInfo as part of the launch intent,
compare that version with the app's version - if the update isn't newer
then disable the update button and change the displayed text.

To make implementation cleaner, move the private currentVersion function
from the UpdateNotifier class to the companion object of the Version
class.

Also add an androidTest class for UpdateActivity to check for the new
behaviour.  This also required setting max-depth: 0 on the integration
jobs so that git describe can find a tag (and improve the
IllegalArgumentException text in the Version ctor to include the
unexpected value which helped to debug the issue).