Add a timeout input to the dbclient action
Enhance the dbclient action with a timeout input which when above zero
then waits for this number of seconds before killing the dbclient
process.  This should allow users to more easily debug issues like #118
(where dbclient hangs indefinitely) by returning the output so far, as
opposed to the current behaviour that the output is lost when Tasker
times out the plugin action.

As the standard error and standard output streams end up being closed
immediately when the process is killed, read these in background threads
instead of after the process has exited so that as much output as
possible is captured.

Add tests for the new behaviour, plus for correctly returning the
content of stderr and success on an exit code of zero (as these were
previously missing test coverage and ended up being broken during the
implementation).