Wait for end of output streams in DbclientRunner
A side-effect of moving the dbclient stderr and stdout stream reading
into separate threads is that the output lines might still be being read
after the process exits.  Add some flags and a wait loop to
DbclientRunner to prevent this (as well as an abort flag to prevent a
hang after killing the process)

Also improve the captureStdout test to check a line near the end of the
help output instead of both from near the start.