Disable OldTargetApi lint check for the moment
API level 34 changes foreground services to require that a service type
is specified.  As taskerpluginlibrary launches a foreground service to
prevent the plugin activity being terminated after a short time it seems
likely that an update will be required first (or this will need to be
confirmed on a device running Android 14).