Update targetSdk to 34
The only change obviously affecting the app in this version relates to
foreground service types becoming mandatory but the Tasker plugin
library has already been updated to handle this for us.

Keep the OldTargetApi link check disabled but update the comment to
reflect that GitHub installing pre-release SDKs on the Actions runners
causes this to trigger before the stable versions have been released.