Return errors from runners instead of throwing
Now that TaskerPluginResultErrorWithOutput has been added to the Tasker
Plugin Library it is possible to return an error result from an action
which returns data without having to resort to throwing an exception.

Update the relevant tests and change the behaviour of the public key,
Dropbear and rsync actions to return errors for 'expected' issues
encountered.  As TaskerPluginResultErrorWithOutput takes an error
code (Tasker sets %err to this value) as well as an error message,
consistently set this to either the command exit code or 100 when the
error is caused by a missing private key.

Also extend some of the existing tests to check that the content of
stderr is returned as the error message if a non-zero exit code is the
reason for returning.  Add a new 'errorFromFailure' test for the public
key action to check non-zero exit is returned as an error.