Add auto process kill before timeout for rsync too
Extract the logic for killing the dbclient process if the action timeout
is near (without losing the content of the stdout and stderr streams)
into a new ProcessHandler class.  Use this new class to extend the
functionality to the rsync action too.

Adjust the way that a timed-out process is destroyed on older devices.
In API 25 and below, Process.destroy() sends SIGKILL instead of SIGTERM
to the process in question.  This doesn't matter for dbclient but with
rsync it means that the cleanup code in the signal handler doesn't get
run, meaning that the destroy() call itself hangs when trying to close
the output streams.  Work around this by parsing the process pid out of
.toString() output if possible and using Process.sendSignal to send
SIGTERM ourselves.  Also change the Process.destroyForcibly() call for
API 26 and above to Process.destroy() - currently these both send
SIGTERM but if the Process-related classes in Android are synced with
upstream OpenJDK at some point in the future destroyForcibly could start
to send SIGKILL instead.

Also take the opportunity to fix a race condition on API 25 and below
which occasionally prevented the timeout logic from ending the process.