Refactor away usage of Compat notification classes
Replace usage of the androidx backward compatability classes in
UpdateNotifier with the standard Android classes, using
Build.VERSION.SDK_INT to determine exactly which calls to use.

This will make it much clearer in future when minSdk is increased that
particular parts of the code are for specified API levels as opposed to
the backward compatability classes which wouldn't be immediately obvious
they weren't needed.