Clean up pubkey file written by Dropbear 2024.84
As dropbearkey (and so also the new dropbearkey.pub) are being written
to a location that the user can't access (and we extract the public key
on demand from the private key) it is cleanest to remove the .pub file.

Add a test for dropbearkey.pub being absent after running the private
key action and adjust the action itself to perform the deletion.