Add explanation for perm overlooked in 6fe321adad
The FOREGROUND_SERVICE_SPECIAL_USE permission isn't declared in
AndroidManifest.xml but comes from the taskerpluginlibrary dependency.
Add this to README.md (and also fix the permission list indent to match
the actions list above).

To ensure the list in README.md always matches the app in future, add a
new 'readmePermsCheck' task to the app build script.  This reads the
permissions in the merged manifest, extracts those listed in README.md
and compares the two, throwing an exception if they differ.  Also set
this new task as a dependency of the 'check' task so that it fails the
GradleCheck Overcommit hook and the CI Build job.

Also remove the FOREGROUND_SERVICE permission from AndroidManifest.xml
as this too is declared by taskerpluginlibrary (and is an implementation
detail).

Resolves #174.