Default update check off if installed from a store
Make the checkForUpdates field on the RsyncConfig and DbclientConfig
classes nullable and default it to null instead of true.  When setting
the toggle value in assignFromInput from the config activities, use the
chosen value unless it is null.  If the value is null and
getInstallSourceInfo.packageSource returns PACKAGE_SOURCE_STORE, set it
to off as the store would be expected to be managing update checks
itself.  Otherwise (or below API level 33 where this is not available),
continue setting it to on.

This is not very well tested but it doesn't seem at all evident how to
set the install source for a package installed via adb.  However, as in
all other cases the behaviour is unchanged it seems worth trying.