Clearer error when missing perm will affect rsync
Instead of an rsync action failing with a file access error (or worse,
being able to read only apparently empty directories!), add an extra
check before launching rsync.  If MANAGE_EXTERNAL_STORAGE or
WRITE_EXTERNAL_STORAGE (depending on Android version) hasn't been
granted, try and determine if the source, destination, or input file
paths are located on 'external storage' local to the device and if so,
return a clearer error message.

To assist with extracting relevant paths from the rsync arguments, add a
new RsyncArgExtractor class and matching tests.  For further validation,
add new missing permission integration tests for the rsync source dir,
dest dir or log file, plus an extra test for the same behaviour when
using the old /sdcard symlink instead of the actual path.

As removing the MANAGE_EXTERNAL_STORAGE permission causes the app to be
killed by Android (interrupting the tests), change the RsyncRunnerTest
test method order to alphabetical and rename copyFromAndToPrimaryStorage
to syncFromAndToPrimaryStorage so that it sorts after the new
errorMessageFromMissing* tests.

Also adjust the util function which was previously used for granting
MANAGE_EXTERNAL_STORAGE so that it is safe to call with enabled: false
(if permission is not already granted) and to also manage the
WRITE_EXTERNAL_STORAGE permission for older versions of Android

Added a missing assert for syncFromAndToPrimaryStorage success noticed
during testing.

Resolves #49.