Add Overcommit check for CI API level coverage
Add a new 'TestedApiLevels' Overcommit check which extracts the minSdk
and targetSdk values from build.gradle.kts and checks that both values
are included in the matrix.

Add API Level 34 to the matrix to satisfy this check and drop the 21
entry back to 19 as the issue encountered in 03c1f25e2e is now resolved.
As x86 images no-longer appear to be made available for API Level 31 and
above, set arch to x86_64 instead when needed.