Drop Dependabot PR limits from five to three
As the build followed by integration tests takes over five minutes to
run it can be quite tedious to wait for multiple remaining PRs to rebase
and pass CI before they can be merged.  Reduce the open PR limits from
the default (five) to three so that larger batches of releases are
spread over multiple days.