Free up disk space to fix CI failure for API 34
The instrumented tests for API level 34 have just started to fail with
the error "Not enough space to create userdata partition" (which goes on
to say that 7372 megabytes are needed but only 7369 are available).
Delete everything under /usr/lib/mono to free up a couple of hundred
megabytes more space.