Explicitly set Gradle Java toolchain version to 17
Building with Android Studio 2024.2.1 was generating a warning as the
bundled JDK (21) has deprecated support for targeting Java 1.8.  Resolve
this by setting the toolchain used by Gradle to JDK 17.

Curiously, it was then possible to remove the sourceCompatibility and
targetCompatibility options for Java compilation as well as the
jvmTarget option for compiling Kotlin.  Although they default to the
configured toolchain version (i.e. 17) it appears that the AGP build
process can convert this to bytecode suitable for the minSdk of 21.