Bump compileSdk to fix build with core-ktx 1.15
core-ktx 1.15.0 requires a compileSdk of 35 or above so bump this from
34 to 35.

However, this causes the current() function in the Version class
companion object to fail to compile as PackageInfo.versionName can be
null.  Resolve this by enabling the generation of the BuildConfig class
and using VERSION_NAME from it instead as this is known at compile time.

Also take the opportunity to convert the current() function into a lazy
property as an argument is no-longer required (and this makes things
cleaner on the calling side).