Remove GPR repository config and token handling
Now that both the android-dropbear and android-rsync packages are
published to Maven Central (because it doesn't require authentication
for package downloads), remove the no-longer required repo-specific
configuration and token passing/handling.

This should get us a step closer with reproducible builds (issue #218),
as the need for an access token was preventing a successful build.