Only require private key for remote rsync src/dest
Currently, the rsync action checks if a private key has been generated
and if not returns an error message telling the user to generate one.
This is not useful when only local paths are being passed to rsync.

Extend RsyncArgExtractor (and add unit tests) to detect when a passed
non-option argument involves a remote path (e.g. contains a colon) and
make the result available as a boolean property.

Use the new property to only check for a private key in RsyncRunner when
a source or destination path is remote.  Adjust the noPrivateKey and
syncFromAndToPrimaryStorage integration tests to correctly test a
missing key with a remote path and a missing key with only local ones.
Also remove creation of a dummy key file from integration tests that
no-longer need it.

Resolves #130.