Enable edge-to-edge mode in all activities
In preparation for bumping targetSdk to 35 (which enables edge-to-edge
mode automatically), enable edge-to-edge mode for the config and update
activities.

To allow adding the necessary padding to the app bar, switch from the
theme automatically providing an action bar to manually defining a
toolbar in each activity.

Add a function to the Utils class to set up window insets listeners for
the action bar and content view as this is the same for all activities.

Also take the opportunity to add a ScrollView around the content in each
activity to handle smaller screens gracefully.