Fix Private Key Config crash on screen rotate
If PrivateKeyConfigActivity is destroyed and re-created, such as when
the screen is rotated or the system theme changes it crashes due to
onItemSelected being called with a null view but it being defined with a
non-null View type.  Resolve this by defining view as nullable.

Also add recreate tests for all of the views so that there is coverage
of this behaviour.