Switch description metadata to markdown
Change from the original imported full description value to a copy of
the relevant content from README.md.  Add a blank line to both before
the actions bullet points for wider markdown parser compatibility.

As ending a line with a backslash is a non-standard extension, switch
this to two spaces in the full description.

To prevent either the full or short description values getting out of
sync with the README, add a new Overcommit hook to check that they match
if any of them are updated.