Increase minSdk to 23 for activity-ktx 1.12
activity-ktx 1.12 has a minSdkVersion of 23:
 - Increase our minSdkVersion from 21 to 23 to fix the build.
 - Remove API level 21/22-specific code.
 - Update the CI api-level matrix.

Having minSdkVersion set to 23 or above also causes AGP to default to
storing native libraries uncompressed and aligned in the apk (so that
they can be loaded directly rather than having to be unpacked to
nativeLibraryDir).  Unfortunately this breaks executing our binaries
(Java can only run binaries that exist as executable files) so set
android.packaging.jniLibs.useLegacyPackaging to true to fix this.