Fix integration test perms for API levels 23 to 28
Although the compiler and lint don't flag it as an issue,
UiAutomation.grantRuntimePermission and revokeRuntimePermission don't
exist at all on API level 23 (Android 6) and is only available as a
hidden function in API levels 24 to 27 (Android 7.0 to 8.1).  Work
around this with some more involved reflection for API level 23 and some
straight-forward reflection for API levels 24 to 27.

The syncFromAndToPrimaryStorage test was also failing for API levels 23
to 28 as the WRITE_EXTERNAL_STORAGE permission doesn't appear to imply
READ_EXTERNAL_STORAGE for these.  Fix this by setting the read
permission explicitly too.