Filter unsupported flags warning on Android 6 & 7
The (harmless) warning "unsupported flags DT_FLAGS_1=0x8000001" is
written to stderr by the dynamic linker under Android 6 and 7 when
running the binaries as it does not expect the DF_1_PIE flag to be set
(ref ribbons/android-rsync#39).

This breaks several tests that rely on the content of stderr.  Resolve
this issue by filtering out the line in ProcessHandler when running on
Android 7 and below (and also switching PublicKeyRunner to use
ProcessHandler so it can benefit from the filtering too).